FIX : 반복 규칙 수정/삭제 오류 수정

+++ src/main/java/com/ideality/coreflow/calendar/command/application/service/CalendarCommandFacadeService.java
@@ -63,7 +63,10 @@ public class CalendarCommandFacadeService {
 		}
 
 		calendarCommandService.deletePersonalSchedule(scheduleId);
-		repeatRuleCommandService.deleteRepeatRule(scheduleId);
+		Boolean isRepeatSchedule = calendarCommandService.isRepeatSchedule(scheduleId);
+		if(isRepeatSchedule){
+			repeatRuleCommandService.deleteRepeatRule(scheduleId);
+		}
 
 	}
 }
